refactor: extract Settings page into 3 components (#986)
Split the monolithic tldw.tsx (2,094 lines) into focused components:

- TldwConnectionSettings.tsx (368 lines) — Server URL, auth mode,
  credentials, connection test, health status display
- TldwTimeoutSettings.tsx (387 lines) — Timeout presets, 8 individual
  timeout fields, reset button
- TldwBillingSettings.tsx (645 lines) — Plans, usage, subscription
  management, invoice history with all billing types and formatters

Parent tldw.tsx reduced to 915 lines (56% reduction). Retains all
state declarations, async handlers, and effects. Renders children
via props — Connection + Timeout inside shared Form, Billing outside.

No functionality changes. All i18n keys, classNames, and data-testid
attributes preserved.

Closes #986
